sched/core: Forbid Unity-based games from changing their CPU affinity
Unity-based games (such as Wild Rift) like to shoot themselves in the foot by
 setting a nonsense CPU affinity, restricting the game to a narrow set of CPU
cores that it thinks are the "big" cores in a heterogeneous CPU. It assumes that
CPUs only have two performance domains (clusters), and therefore royally mucks up
games' CPU affinities on CPUs which have more than two performance domains.

Check if a setaffinity target task is part of a Unity-based game and
silently ignore the setaffinity request so that it can't sabotage itself.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>